csp/requirements.txt
flask
requests
cryptography
PyJWT